Moved History from modules to ActionWindow folder.
