+++ src/parser/core/modules/ActionWindow/evaluators/ExpectedActionsEvaluator.tsx
@@ -2,9 +2,9 @@ import {Plural, Trans} from '@lingui/react'
 import {ActionLink} from 'components/ui/DbLink'
 import {RotationTargetOutcome} from 'components/ui/RotationTable'
 import React from 'react'
-import {HistoryEntry} from '../../History'
 import {SeverityTiers, TieredSuggestion} from '../../Suggestions/Suggestion'
 import {EvaluatedAction} from '../EvaluatedAction'
+import {HistoryEntry} from '../History'
 import {OutcomeCalculator, TrackedAction, TrackedActionsOptions} from './TrackedAction'
 import {EvaluationOutput, WindowEvaluator} from './WindowEvaluator'
