+++ src/parser/core/modules/ActionWindow/evaluators/LimitedActionsEvaluator.tsx
@@ -1,8 +1,8 @@
 import {Plural, Trans} from '@lingui/react'
 import React from 'react'
-import {HistoryEntry} from '../../History'
 import {SeverityTiers, TieredSuggestion} from '../../Suggestions/Suggestion'
 import {EvaluatedAction} from '../EvaluatedAction'
+import {HistoryEntry} from '../History'
 import {TrackedAction, TrackedActionsOptions} from './TrackedAction'
 import {WindowEvaluator} from './WindowEvaluator'
