+++ src/parser/jobs/brd/modules/RagingStrikes.tsx
@@ -11,10 +11,10 @@ import {EvaluatedAction} from 'parser/core/modules/ActionWindow/EvaluatedAction'
 import {ExpectedActionsEvaluator} from 'parser/core/modules/ActionWindow/evaluators/ExpectedActionsEvaluator'
 import {ExpectedGcdCountEvaluator} from 'parser/core/modules/ActionWindow/evaluators/ExpectedGcdCountEvaluator'
 import {TrackedAction, TrackedActionsOptions} from 'parser/core/modules/ActionWindow/evaluators/TrackedAction'
+import {HistoryEntry} from 'parser/core/modules/ActionWindow/History'
 import {BuffWindow} from 'parser/core/modules/ActionWindow/windows/BuffWindow'
 import {Actors} from 'parser/core/modules/Actors'
 import {GlobalCooldown} from 'parser/core/modules/GlobalCooldown'
-import {HistoryEntry} from 'parser/core/modules/History'
 import {SEVERITY} from 'parser/core/modules/Suggestions'
 import React from 'react'
 import {Team} from 'report'
