+++ src/parser/jobs/drg/modules/DragonSight.tsx
@@ -8,9 +8,9 @@ import {EvaluatedAction} from 'parser/core/modules/ActionWindow/EvaluatedAction'
 import {ExpectedActionsEvaluator} from 'parser/core/modules/ActionWindow/evaluators/ExpectedActionsEvaluator'
 import {calculateExpectedGcdsForTime, ExpectedGcdCountEvaluator} from 'parser/core/modules/ActionWindow/evaluators/ExpectedGcdCountEvaluator'
 import {EvaluationOutput, WindowEvaluator} from 'parser/core/modules/ActionWindow/evaluators/WindowEvaluator'
+import {HistoryEntry} from 'parser/core/modules/ActionWindow/History'
 import {BuffWindow} from 'parser/core/modules/ActionWindow/windows/BuffWindow'
 import {GlobalCooldown} from 'parser/core/modules/GlobalCooldown'
-import {HistoryEntry} from 'parser/core/modules/History'
 import {SEVERITY} from 'parser/core/modules/Suggestions'
 import React from 'react'
 import {isDefined} from 'utilities'
