+++ src/parser/jobs/gnb/modules/NoMercy.tsx
@@ -7,9 +7,9 @@ import {ExpectedActionsEvaluator} from 'parser/core/modules/ActionWindow/evaluat
 import {ExpectedGcdCountEvaluator} from 'parser/core/modules/ActionWindow/evaluators/ExpectedGcdCountEvaluator'
 import {NotesEvaluator} from 'parser/core/modules/ActionWindow/evaluators/NotesEvaluator'
 import {TrackedAction} from 'parser/core/modules/ActionWindow/evaluators/TrackedAction'
+import {HistoryEntry} from 'parser/core/modules/ActionWindow/History'
 import {BuffWindow} from 'parser/core/modules/ActionWindow/windows/BuffWindow'
 import {GlobalCooldown} from 'parser/core/modules/GlobalCooldown'
-import {HistoryEntry} from 'parser/core/modules/History'
 import {SEVERITY} from 'parser/core/modules/Suggestions'
 import React from 'react'
