+++ src/parser/jobs/mch/modules/Tincture.tsx
@@ -5,8 +5,8 @@ import {dependency} from 'parser/core/Module'
 import {EvaluatedAction} from 'parser/core/modules/ActionWindow/EvaluatedAction'
 import {ExpectedActionsEvaluator} from 'parser/core/modules/ActionWindow/evaluators/ExpectedActionsEvaluator'
 import {TrackedAction, TrackedActionsOptions} from 'parser/core/modules/ActionWindow/evaluators/TrackedAction'
+import {HistoryEntry} from 'parser/core/modules/ActionWindow/History'
 import Downtime from 'parser/core/modules/Downtime'
-import {HistoryEntry} from 'parser/core/modules/History'
 import {SEVERITY} from 'parser/core/modules/Suggestions'
 import {Tincture as CoreTincture} from 'parser/core/modules/Tincture'
 import React from 'react'
