+++ src/parser/jobs/sam/modules/Meikyo.tsx
@@ -5,9 +5,9 @@ import {dependency} from 'parser/core/Injectable'
 import {EvaluatedAction} from 'parser/core/modules/ActionWindow/EvaluatedAction'
 import {AllowedGcdsOnlyEvaluator} from 'parser/core/modules/ActionWindow/evaluators/AllowedGcdsOnlyEvaluator'
 import {ExpectedGcdCountEvaluator} from 'parser/core/modules/ActionWindow/evaluators/ExpectedGcdCountEvaluator'
+import {HistoryEntry} from 'parser/core/modules/ActionWindow/History'
 import {BuffWindow} from 'parser/core/modules/ActionWindow/windows/BuffWindow'
 import {GlobalCooldown} from 'parser/core/modules/GlobalCooldown'
-import {HistoryEntry} from 'parser/core/modules/History'
 import {SEVERITY} from 'parser/core/modules/Suggestions'
 import React from 'react'
 import DISPLAY_ORDER from './DISPLAY_ORDER'
